refactor: move makeInterruptibleAsyncInterval to monitor.ts
